Refactored GeoModel.
Changed code in GeoModel to be more "understanding" of GeoJSON
structures. It should now produce reasonable results for any valid
GeoJSON data.

Changed the `extractGeometryFeatures()` method to set the geometry
feature field to a valid GeoJSON object, which is what is needed for the
MongoDB `2dsphere` index.

Removed `geometrySingle` property from GeoModel. It wasn't used, and
didn't seem to serve a useful purpose. Same for the
`extractGeometryFeatures` property.

Removed the `query` parameter from the `findWithin()` and
`findIntersects()` methods. It wasn't used, and the code implementing it
wasn't correct anyway.

Added `getGeometry()` and `getCentroid()` methods to extract geometry
data from the GeoJSON geographic field of a GeoModel instance. They're
intended use is for extracting coordinate data that can be used for geo
queries. The `getGeometry()` method extracts GeoJSON geometry objects.
The `getCentroid()` method calculates the centroid of the geographic
field.

Documented the code.

Added tests to check functionality of GeoModel. Wired it up to a
`sails-mongo` adapter so it could actually exercise MongoDB geo
functionality.