Work around Waterline bug with .where() clause applied to empty query.
The GeoModel `.findWithin()` and `.findIntersects()` methods return a
query consisting of match criterion of "id" to an array of values. The
query can be refined with additional method calls. However, if the id
array is empty, Waterline appears to ignore it when a `.where()` clause
is applied to the query. (Works fine without the `.where()` clause.)

Worked around the problem by replacing the empty array with an empty
string (which should never match a valid id). Yuck!

Added a test for this case.