Documentation changes and minor code cleanup.
Reorganized documentation generation to use a `documentation.yml`
configuration file. Also, lots of changes, additions and corrections to
documentation.

The `BaseModel#extend()` method accepted a second parameter for
specifying static properies of the model definition. The Waterline
functionality it used has been deprecated, and it appears to have never
been used. So I removed it.

The storage configuration `modelsDir` property seems to not be used. (A
leftover from earlier versions or a never-implemented feature, perhaps?)
Removed some references to it.

In src/test/GeoModel.js, imported bluebird (needed for
`Promise.mapSeries()`).