Removed default adapters and connections definitions.
Motivated by need to get rid of waterline-sqlite3 dependency, which is
not well supported on node v10 or v12.

Instead of trying to switch the defaults to something else (e.g.
sails-mongo), decided it was simpler to remove the defaults. This seems
pretty safe for `adapters`, since the `waterline-sqlite3` default wasn't
useful, and seems to have been universally redefined. The `connections`
default also seems to be univerally redefined, because of the need to
define the `url` attribute.