Fix to emit destroy event for each model instance destroyed.
In the current code, when the `destroy()` method is invoked to destroy
multiple model instances, destroy events are emitted for only the first
of the instances. In the updated code, events are emitted for each of
the instances.

(These missing events are a major problem for nxus-searcher, causing it
to leave dangling ElasticSearch records.)

(Is it reasonable to emit multiple events like this?)